general: make Buffer::get_vec() and get_or_none() generic
So the tedious `.. as usize` can be avoided.

+++ src/table/bitmap/ebdt.rs
@@ -54,7 +54,7 @@ impl Font {
                                 + 1;
                             (0..len).for_each(|_| {
                                 strike_bitmap_data.push(BitmapData {
-                                    image_data: Some(buffer.get_vec(image_size as usize)),
+                                    image_data: Some(buffer.get_vec(image_size)),
                                     ..Default::default()
                                 });
                             })
